Lazy load font faces with JS
add "fontfaces" id to noscript tags with font face stylesheets.
Modify webpack config to generate two bundles: one for home page and
the other for posts.
Add new js module fonts.js which looks for the noscript element with
fontfaces id and takes the url of the contained link to load the the
fontface stylesheets after the document loads.